Extract commander actions into an actions folder
